requirements.txt
asgiref==3.5.2
Babel==2.10.3
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-phone-field==1.8.1
django-phonenumber-field==7.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.3
phonenumbers==8.12.55
phonenumberslite==8.12.55
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
